+++ docs/gallery/gallery/layout/nested_layouts.jl
@@ -1,7 +1,7 @@
 # ---
 # title: Nested layouts
 # cover: assets/nested_layouts.png
-# description: Embedding AlgebraOfGraphics plots in a Makie figure
+# description: AlgebraOfGraphics plots within a Makie figure.
 # author: "[Pietro Vertechi](https://github.com/piever)"
 # ---
 
@@ -12,8 +12,7 @@ set_aog_theme!() #src
 # of the figure is managed by vanilla Makie.
 # For example
 
-resolution = (800, 600)
-fig = Figure(; resolution)
+fig = Figure(; size=(800, 600))
 ax = Axis(fig[1, 1], title="Some plot")
 
 df = (
@@ -33,6 +32,38 @@ for ae in ag
 end
 legend!(fig[end+1, 2], ag, orientation=:horizontal, tellheight=true)
 fig
+
+# The above also works in more nested situations.
+
+f = Figure(; size=(800, 600))
+ax = Axis(f[1, 1], title="Some plot")
+subfig = f[1, 2]
+ax2 = Axis(subfig[1, 1])
+
+df = (
+    x=rand(500),
+    y=rand(500),
+    c=rand(["a", "b", "c"], 500),
+)
+plt = data(df) * mapping(:x, :y, color=:c)
+
+draw!(subfig[2, 1], plt)
+f
+
+# It is also possible to let Makie control the axis and plot directly on top of it.
+
+f = Figure(; size=(800, 300))
+ax1 = Axis(f[1, 1])
+ax2 = Axis(f[1, 2])
+
+df = (x=rand(100), y=rand(100), c=rand(["a", "b", "c"], 100))
+plt = data(df) * mapping(:x, :y, color=:c)
+
+scatter!(ax1, rand(10), rand(10), color=:black)
+grid = draw!(ax2, plt)
+legend!(f[1, 3], grid)
+
+f
 
 # save cover image #src
 mkpath("assets") #src