+++ src/entries.jl
@@ -1,10 +1,12 @@
+const PlotType = Type{<:Plot}
+
 """
-    Entry(plottype::PlotFunc, positional::Arguments, named::NamedArguments)
+    Entry(plottype::PlotType, positional::Arguments, named::NamedArguments)
 
 Define plottype as well as positional and named arguments for a single plot.
 """
 struct Entry
-    plottype::PlotFunc
+    plottype::PlotType
     positional::Arguments
     named::NamedArguments
 end
@@ -24,7 +26,7 @@ function Base.append!(e1::Entry, e2::Entry)
     return Entry(plottype, positional, named)
 end
 
-# Use technique from https://github.com/JuliaPlots/AlgebraOfGraphics.jl/pull/289
+# Use technique from https://github.com/MakieOrg/AlgebraOfGraphics.jl/pull/289
 # to encode all axis information without creating the axis.
 struct AxisSpec
     type::Union{Type{Axis}, Type{Axis3}}
@@ -42,8 +44,8 @@ end
 struct AxisSpecEntries
     axis::AxisSpec
     entries::Vector{Entry}
-    categoricalscales::MixedArguments
-    continuousscales::MixedArguments
+    categoricalscales::Dictionary{KeyType, CategoricalScale}
+    continuousscales::Dictionary{KeyType, ContinuousScale}
 end
 
 """
@@ -56,8 +58,8 @@ scale should be a `ContinuousScale`.
 struct AxisEntries
     axis::Union{Axis, Axis3}
     entries::Vector{Entry}
-    categoricalscales::MixedArguments
-    continuousscales::MixedArguments
+    categoricalscales::Dictionary{KeyType, CategoricalScale}
+    continuousscales::Dictionary{KeyType, ContinuousScale}
 end
 
 function AxisEntries(ae::AxisSpecEntries, fig)
@@ -65,17 +67,15 @@ function AxisEntries(ae::AxisSpecEntries, fig)
     AxisEntries(ax, ae.entries, ae.categoricalscales, ae.continuousscales)
 end
 
-function AxisEntries(ae::AxisSpecEntries, ax::Union{Axis,Axis3})
-    if !isempty(ax)
-        @warn("Axis got passed, but also axis attributes. Ignoring axis attributes $(a.axis.attributes)")
-    end
+function AxisEntries(ae::AxisSpecEntries, ax::Union{Axis, Axis3})
     AxisEntries(ax, ae.entries, ae.categoricalscales, ae.continuousscales)
 end
 
 function Makie.plot!(ae::AxisEntries)
     axis, entries = ae.axis, ae.entries
     for entry in entries
-        plot!(entry.plottype, axis, entry.positional...; pairs(entry.named)...)
+        plot = entry.plottype(Tuple(entry.positional), Dict{Symbol, Any}(pairs(entry.named)))
+        plot!(axis, plot)
     end
     return ae
 end
@@ -104,7 +104,5 @@ function FileIO.save(file::FileIO.Formatted, fg::FigureGrid; kwargs...)
     return FileIO.save(file, fg.figure; kwargs...)
 end
 
-to_tuple(fg) = (fg.figure, fg.grid)
-
-Base.iterate(fg::FigureGrid) = iterate(to_tuple(fg))
-Base.iterate(fg::FigureGrid, i) = iterate(to_tuple(fg), i)
+Base.iterate(fg::FigureGrid) = iterate((fg.figure, fg.grid))
+Base.iterate(fg::FigureGrid, i) = iterate((fg.figure, fg.grid), i)