+++ src/transformations/linear.jl
@@ -2,9 +2,15 @@ Base.@kwdef struct LinearAnalysis{I}
     npoints::Int=200
     dropcollinear::Bool=false
     interval::I=automatic
+    level::Float64=0.95
 end
 
-add_intercept_column(x::AbstractVector{T}) where {T} = [ones(T, length(x)) x]
+function add_intercept_column(x::AbstractVector{T}) where {T}
+    mat = similar(x, float(T), (length(x), 2))
+    fill!(view(mat, :, 1), 1)
+    copyto!(view(mat, :, 2), x)
+    return mat
+end
 
 # TODO: add multidimensional version
 function (l::LinearAnalysis)(input::ProcessedLayer)
@@ -14,14 +20,14 @@ function (l::LinearAnalysis)(input::ProcessedLayer)
         default_interval = length(weights) > 0 ? nothing : :confidence
         interval = l.interval === automatic ? default_interval : l.interval
         # FIXME: handle collinear case gracefully
-        lin_model = GLM.lm(add_intercept_column(x), collect(y); wts=weights, l.dropcollinear)
+        lin_model = GLM.lm(add_intercept_column(x), y; wts=weights, l.dropcollinear)
         x̂ = range(extrema(x)..., length=l.npoints)
-        pred = GLM.predict(lin_model, add_intercept_column(x̂); interval)
+        pred = GLM.predict(lin_model, add_intercept_column(x̂); interval, l.level)
         return if !isnothing(interval)
-            ŷ, lower, upper = map(vec, pred) # GLM prediction returns matrices
+            ŷ, lower, upper = pred
             (x̂, ŷ), (; lower, upper)
         else
-            ŷ = vec(pred) # GLM prediction returns matrix
+            ŷ = pred
             (x̂, ŷ), (;)
         end
     end
@@ -31,14 +37,18 @@ function (l::LinearAnalysis)(input::ProcessedLayer)
 end
 
 """
-    linear(; interval=automatic, dropcollinear=false, npoints=200)
+    linear(; interval=automatic, level=0.95, dropcollinear=false, npoints=200)
 
 Compute a linear fit of `y ~ 1 + x`. An optional named mapping `weights` determines the weights.
-Use `interval` to specify what type of interval the shaded band should represent.
-Valid values of interval are `:confidence` delimiting the uncertainty of the predicted
-relationship, and `:prediction` delimiting estimated bounds for new data points.
+Use `interval` to specify what type of interval the shaded band should represent,
+for a given coverage `level` (the default `0.95` equates `alpha = 0.05`).
+Valid values of `interval` are `:confidence`, to delimit the uncertainty of the predicted
+relationship, and `:prediction`, to delimit estimated bounds for new data points.
+Use `interval = nothing` to only compute the line fit, without any uncertainty estimate. 
 By default, this analysis errors on singular (collinear) data. To avoid that,
 it is possible to set `dropcollinear=true`.
-`npoints` is the number of points used by Makie to draw the line
+`npoints` is the number of points used by Makie to draw the shaded band.
+
+Weighted data is supported via the keyword `weights` (passed to `mapping`).
 """
 linear(; options...) = transformation(LinearAnalysis(; options...))